fz.ptvo_multistate_action, fz.ptvo_switch_uart and fz.ptvo_switch_analog_input,
